requirements.txt
opencv-python
scikit-learn
scikit-image
scikit-fuzzy
pandas
numpy
matplotlib
pillow
fuzzywuzzy
python-levenshtein
requests
folium
plotly
tensorflow
pygments
tk
tkhtmlview
